% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_fop.R
\name{install_fop}
\alias{install_fop}
\title{Install or Update the Apache FOP jar}
\usage{
install_fop(version = "2.3",
  digest = "1bc3b660e5bbc470fbc9e4227fc55ac52a8624b9f5efac1d68bc693f044ad095d37286ebb5d7242d9a01e8f1c8b62d77d2c8803914eb48970e536499b6a2053a",
  mirrors = c("http://mirrors.ocf.berkeley.edu/apache/xmlgraphics/fop/",
  "http://apache.cs.utah.edu/xmlgraphics/fop/",
  "http://mirror.cc.columbia.edu/pub/software/apache/xmlgraphics/fop/"),
  url = character())
}
\arguments{
\item{version}{The declared FOP version}

\item{digest}{The sha15 checksum. Set to an empty string "" to skip the check.}

\item{mirrors}{A vector Apache mirror sites. One is picked randomly.}

\item{url}{Optional url of a particular file to download. Will override downloading from random mirrors.}
}
\value{
Logical if the installation was successful, invisibly.
}
\description{
This downloads and installs the FOP App jar (~4.1 MB) into a directory, and verifies the integrity of the file using a checksum.
}
\details{
This function will download and install the Apache Formatting Objects Processor, FOP. A ZIP archive is saved in the temp directory and then extracted. To find the extracted files run \code{rappdirs::site_data_dir("rDOM", "cscheid")}. The default configuration file is then copied to \code{rappdirs::user_config_dir("rDOM", "cscheid")} and used as the default configuration file for FOP. 

This function borrows heavily from \code{\link[rtika]{install_tika}}.
}
\examples{
\dontrun{
install_fop()
}
}
\seealso{
\url{https://xmlgraphics.apache.org/fop/download.html}
}
