% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{printDocument}
\alias{printDocument}
\title{printDocument}
\usage{
printDocument(doc)
}
\arguments{
\item{doc}{An \code{\link{rJava}} object representing the document.}
}
\description{
Print a DOM document.
}
