% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdom.R
\name{rdom}
\alias{rdom}
\title{Return DOM of a website as HTML}
\usage{
rdom(url = NULL, css = NULL, all = FALSE, timeout = 30)
}
\arguments{
\item{url}{a URL of a website.}

\item{css}{a string containing one or more CSS selectors separated by commas.}

\item{all}{logical. This controls whether \code{querySelector} or
\code{querySelectorAll} is used to extract elements from the page.
When FALSE, output is similar to \code{rvest::html_node}. When TRUE,
output is similar to \code{rvest::html_nodes}.
If \code{css} is missing, then this argument is ignored.}

\item{timeout}{maximum time to wait for page to load and render, in seconds.}
}
\description{
Return DOM of a website as HTML
}
\examples{
library("rvest")
stars <- "http://www.techstars.com/companies/stats/"
# doesn't work
html(stars) \%>\% html_node(".table75") \%>\% html_table()
# should work
rdom(stars) \%>\% html_node(".table75") \%>\% html_table()
# more efficient
stars \%>\% rdom(".table75") \%>\% html_table()
}

