% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{saveDocument}
\alias{saveDocument}
\title{saveDocument}
\usage{
saveDocument(doc, path)
}
\arguments{
\item{doc}{An \code{\link{rJava}} object representing the document.}

\item{path}{File path and name for the resulting file.}
}
\description{
Save a DOM document to a file.
}
