% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rdom.R
\name{rdom}
\alias{rdom}
\title{Return DOM of a website as HTML}
\usage{
rdom(url, timeout = 10)
}
\arguments{
\item{url}{a URL of a website.}

\item{timeout}{maximum time to wait for page to load and render, in seconds}
}
\description{
Return DOM of a website as HTML
}
\examples{
library("rvest")
# doesn't work
html("http://www.techstars.com/companies/stats/") \%>\%
  html_node(".table75") \%>\% html_table()
# should work
rdom("http://www.techstars.com/companies/stats/") \%>\%
  html_node(".table75") \%>\% html_table()
}

